Isolate some fuctions outside Phaser and make test for them, Make sure that input value will be empty after click on save or main buttons
