+++ src/Scenes/HighScoresScene.js
@@ -2,6 +2,7 @@ import Phaser from 'phaser';
 import { getData } from '../ScoreApi';
 import ScrollingBackground from '../Objects/ScrollingBackground';
 
+
 class HighScoresScene extends Phaser.Scene {
   constructor() {
     super({ key: 'HighScores' });
@@ -18,7 +19,6 @@ class HighScoresScene extends Phaser.Scene {
     this.title.setOrigin(0.5);
 
     getData().then((data) => {
-      // console.log(data);
       data.result.sort((a, b) => b.score - a.score)
         .slice(0, 5)
         .map((game, i) => {
